edit done

+++ src/main/java/dream/tk/controller/BusinessFileController.java
@@ -1,10 +1,12 @@
 package dream.tk.controller;
 
 import java.io.File;
+import java.io.IOException;
 
 import javax.servlet.http.HttpSession;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.jdbc.core.SqlOutParameter;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.multipart.MultipartFile;
@@ -36,23 +38,34 @@ public class BusinessFileController {
 	
 	@RequestMapping("editFile")
 	public String editFile(String id, String[] delFiles, MultipartFile[] file) {
+		System.out.println("새롭게 업로드된 파일은? "+file[0].getOriginalFilename());
+		System.out.println("새롭게 업로드된 파일은? "+file[0].getName());
+		System.out.println("새롭게 업로드된 파일은? "+file[0].getSize());
+		
 		
 		String realPath = session.getServletContext().getRealPath("files");
 		try {
+			if(file[0].getSize()>0) {
 			fser.upload(realPath, id, file);
-	
-		for(int i=0; i<delFiles.length; i++) {
-			String seq = delFiles[i];
-			String delfileName = fser.getSysName(seq);
-			File delTarget = new File(realPath+"/"+delfileName);
-			boolean delResult = delTarget.delete();
-			if(delResult) {
-			fser.delete(seq);
 			}
-		}
 		
+			if(delFiles!=null) {
+				for(int i=0; i<delFiles.length; i++) {
+						//System.out.println("지울 seq "+delFiles[i]);
+						String seq = delFiles[i];
+						String delfileName = fser.getSysName(seq);
+						//System.out.println("지울 파일의 이름은  "+delfileName);
+						File delTarget = new File(realPath+"/"+delfileName);
+						//System.out.println("지울 파일들의 풀  경로는 "+delTarget);
+						boolean delResult = delTarget.delete();
+						//System.out.println("지운 결과는 "+delResult);
+							if(delResult) {
+									fser.delete(seq);
+											}
+						}
+			}
 		
-		}catch (Exception e){		
+			}catch (Exception e){		
 			e.printStackTrace();
 		}
